# docs/swagger.yaml
definitions:
  dao.Car:
    properties:
      id:
        type: string
      make:
        type: string
      model:
        type: string
    type: object
host: localhost:8080/
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a simple http server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Simple HTTTP Server
  version: "1.0"
paths:
  /car/create:
    post:
      consumes:
      - application/json
      description: Creates a new car when a dao.Car is given
      parameters:
      - description: Add car
        in: body
        name: car
        required: true
        schema:
          $ref: '#/definitions/dao.Car'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dao.Car'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dao.Car'
      summary: Create a car
  /car/delete:
    post:
      consumes:
      - application/json
      description: Deletes a car when a dao.Car.Id is given
      parameters:
      - description: Delete a car
        in: body
        name: car
        required: true
        schema:
          $ref: '#/definitions/dao.Car'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Deletes a car
  /car/get:
    get:
      consumes:
      - application/json
      description: Get all the cars
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dao.Car'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all cars
  /car/get/{id}:
    get:
      consumes:
      - application/json
      description: Get a car when an Id is given
      parameters:
      - description: Car ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dao.Car'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get a car
  /car/update:
    post:
      consumes:
      - application/json
      description: Updates a car when a dao.Car is given
      parameters:
      - description: Update car
        in: body
        name: car
        required: true
        schema:
          $ref: '#/definitions/dao.Car'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dao.Car'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dao.Car'
      summary: Updates a car
swagger: "2.0"